:recycle: refactor: 테스트 중복 코드 클래스 분리
